feat: refactor ChatUser model to a dedicated user package and update imports across the codebase
