+++ surf-chat-api/src/main/kotlin/dev/slne/surf/chat/api/SurfChatApi.kt
@@ -1,7 +1,7 @@
 package dev.slne.surf.chat.api
 
 import dev.slne.surf.chat.api.channel.Channel
-import dev.slne.surf.chat.api.model.ChatUser
+import dev.slne.surf.chat.api.user.ChatUser
 import dev.slne.surf.chat.api.type.MessageType
 import dev.slne.surf.surfapi.core.api.util.requiredService
 import net.kyori.adventure.text.Component
