+++ surf-chat-api/src/main/kotlin/dev/slne/surf/chat/api/model/ChatFormat.kt
@@ -1,5 +1,6 @@
 package dev.slne.surf.chat.api.model
 
+import dev.slne.surf.chat.api.user.ChatUser
 import dev.slne.surf.chat.api.type.MessageType
 import net.kyori.adventure.text.Component
 import java.util.*
