+++ surf-chat-api/src/main/kotlin/dev/slne/surf/chat/api/model/MessageValidator.kt
@@ -1,5 +1,6 @@
 package dev.slne.surf.chat.api.model
 
+import dev.slne.surf.chat.api.user.ChatUser
 import dev.slne.surf.chat.api.type.MessageType
 import dev.slne.surf.chat.api.type.MessageValidationResult
 import net.kyori.adventure.text.Component
