+++ surf-chat-api/src/main/kotlin/dev/slne/surf/chat/api/user/ChatUserSettings.kt
@@ -1,0 +1,8 @@
+package dev.slne.surf.chat.api.user
+
+interface ChatUserSettings {
+    var pmEnabled: Boolean
+    var pingsEnabled: Boolean
+    var pmFriendBypassEnabled: Boolean
+    var channelInvitesEnabled: Boolean
+}
