+++ surf-chat-core/src/main/kotlin/dev/slne/surf/chat/core/service/DatabaseService.kt
@@ -1,7 +1,7 @@
 package dev.slne.surf.chat.core.service
 
 import dev.slne.surf.chat.api.model.DenyListEntry
-import dev.slne.surf.chat.api.model.ChatUser
+import dev.slne.surf.chat.api.user.ChatUser
 import dev.slne.surf.chat.api.model.HistoryEntry
 import dev.slne.surf.surfapi.core.api.util.requiredService
 import it.unimi.dsi.fastutil.objects.ObjectList
@@ -24,7 +24,6 @@ interface DatabaseService {
         type: String? = null,
         rangeMillis: Long? = null,
         message: String? = null,
-        deleted: Boolean? = null,
         deletedBy: String? = null,
         server: String? = null
     ): ObjectList<HistoryEntry>
@@ -36,8 +35,6 @@ interface DatabaseService {
     suspend fun insertHistoryEntry(user: UUID, entry: HistoryEntry)
 
     suspend fun saveAll()
-
-    fun getName(uuid: UUID): String
 
     companion object {
         val INSTANCE = requiredService<DatabaseService>()
