+++ surf-chat-core/src/main/kotlin/dev/slne/surf/chat/core/service/FilterService.kt
@@ -1,6 +1,6 @@
 package dev.slne.surf.chat.core.service
 
-import dev.slne.surf.chat.api.model.ChatUser
+import dev.slne.surf.chat.api.user.ChatUser
 import dev.slne.surf.chat.api.type.MessageValidationResult
 import dev.slne.surf.surfapi.core.api.util.requiredService
 import net.kyori.adventure.text.Component
