+++ surf-chat-fallback/src/main/kotlin/dev/slne/surf/chat/fallback/model/FallbackChatFormat.kt
@@ -1,7 +1,7 @@
 package dev.slne.surf.chat.fallback.model
 
 import dev.slne.surf.chat.api.model.ChatFormat
-import dev.slne.surf.chat.api.model.ChatUser
+import dev.slne.surf.chat.api.user.ChatUser
 import dev.slne.surf.chat.api.type.MessageType
 import net.kyori.adventure.text.Component
 import java.util.UUID
