+++ surf-chat-fallback/src/main/kotlin/dev/slne/surf/chat/fallback/model/entry/FallbackDenylistEntry.kt
@@ -4,7 +4,7 @@ import dev.slne.surf.chat.api.model.DenyListEntry
 
 data class FallbackDenylistEntry(
     override val word: String,
-    override val reason: String,
+    override val reason: String?,
     override val addedAt: Long,
     override val addedBy: String
 ) : DenyListEntry  {
