+++ surf-chat-fallback/src/main/kotlin/dev/slne/surf/chat/fallback/service/FallbackHistoryService.kt
@@ -1,7 +1,7 @@
 package dev.slne.surf.chat.fallback.service
 
 import com.google.auto.service.AutoService
-import dev.slne.surf.chat.api.model.ChatUser
+import dev.slne.surf.chat.api.user.ChatUser
 import dev.slne.surf.chat.api.model.HistoryEntry
 import dev.slne.surf.chat.api.type.MessageType
 import dev.slne.surf.chat.core.service.HistoryService
